+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Gitaly changelog
 
+UNRELEASED
+
+- Sanitize URLs before logging them
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/624
+
 v0.89.0
 
 - Report original exceptions to Sentry instead of wrapped ones by the exception bridge